Move functions to pkg/utils/utils.go
